Removed old App.tsx
